A major update of the architecture of work with the USE.
The architecture of the bot's work has been changed in connection with the addition of a functional allowing, when querying the USE task by its number parameters (get %task number%). Module db-utils now is universal both for the query by the USE number, and on request by the number of the task.

Changed some stored procedures, the storage structure of users and code refactoring performed.